+++ packages/integration-sdk-runtime/src/__tests__/index.test.ts
@@ -1,10 +1,6 @@
-import { BucketMap, FileSystemGraphObjectStore } from '../index';
+import { FileSystemGraphObjectStore } from '../index';
 
 describe('#storage', () => {
-  test('should expose BucketMap', () => {
-    expect(BucketMap).not.toEqual(undefined);
-  });
-
   test('should expose FileSystemGraphObjectStore', () => {
     expect(FileSystemGraphObjectStore).not.toEqual(undefined);
   });