+++ packages/integration-sdk-runtime/src/storage/index.ts
@@ -1,2 +1,3 @@
 export { GraphObjectStore } from '@jupiterone/integration-sdk-core';
 export * from './FileSystemGraphObjectStore';
+export * from './memory';